// src/components/Scheduler/components/Scheduler/types.ts
import type {
  Config,
  SchedulerData,
  SchedulerItemClickData,
  SchedulerProjectData,
} from '../../types/global'
import type { ParsedDatesRange } from '../../utils/getDatesRange'

export type SchedulerProps = {
  data: SchedulerData
  isLoading?: boolean
  config?: Config
  startDate?: string
  onRangeChange?: (range: ParsedDatesRange) => void
  onTileClick?: (data: SchedulerProjectData) => void
  onFilterData?: () => void
  onClearFilterData?: () => void
  onItemClick?: (data: SchedulerItemClickData) => void
}

export type StyledOutsideWrapperProps = {
  showScroll: boolean
}